Rename identifiers in the new artifact package
